backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Data & ML
numpy>=1.26.0
pandas>=2.2.0
scikit-learn==1.6.1
joblib>=1.3.2

# Validation
pydantic>=2.6.0

# GenAI (Gemini)
google-genai>=0.3.0
google-api-core>=2.17.0

# Environment variables (recommended)
python-dotenv>=1.0.1
